add some changes to bdinfo.py
